\name{.pModify}
\alias{.pModify}
\title{changing title and font size}
\usage{.pModify(p, title="", font.size=12)}
\description{changing ggplot object's title and font size}
\details{interal use, not for user.}
\value{ggplot object}
\author{Guangchuang Yu}
\arguments{\item{p}{ggplot object}
\item{title}{graph title}
\item{font.size}{font size}}
