% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg-utilities.R
\name{append_kegg_category}
\alias{append_kegg_category}
\title{append_kegg_category}
\usage{
append_kegg_category(x)
}
\arguments{
\item{x}{KEGG enrichment result}
}
\value{
update KEGG enrichment result with category information
}
\description{
add KEGG pathway category information
}
\details{
This function appends the KEGG pathway category information to KEGG enrichment result
(either output of 'enrichKEGG' or 'gseKEGG'
}
\author{
Guangchuang Yu
}
