% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitr.R
\name{bitr}
\alias{bitr}
\title{bitr}
\usage{
bitr(geneID, fromType, toType, annoDb, drop = TRUE)
}
\arguments{
\item{geneID}{input gene id}

\item{fromType}{input id type}

\item{toType}{output id type}

\item{annoDb}{annotation db}

\item{drop}{drop NA or not}
}
\value{
data.frame
}
\description{
Biological Id TRanslator
}
\author{
Guangchuang Yu
}

