% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg-utilities.R
\name{browseKEGG}
\alias{browseKEGG}
\title{browseKEGG}
\usage{
browseKEGG(x, pathID)
}
\arguments{
\item{x}{an instance of enrichResult or gseaResult}

\item{pathID}{pathway ID}
}
\value{
url
}
\description{
open KEGG pathway with web browser
}
\author{
Guangchuang Yu
}
