% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_score.R
\name{categ_dist}
\alias{categ_dist}
\title{calculate categorical distance (Cohen's Kappa score) between multiple categories}
\usage{
categ_dist(mapping_table, terms_to_compare = unlist(unique(mapping_table[, 2,
  with = F])), ignore_limit = F)
}
\arguments{
\item{mapping_table}{two-column data.table which provides mapping from object to term}

\item{terms_to_compare}{a character vector specifying which terms to compare}

\item{ignore_limit}{logical, ignore the limit set on the number of terms to compare (1000)}
}
\description{
The categ_dist function to calculate categorical distance (Cohen's Kappa score) between multiple terms (categories).

The function is intended to measure distances between GO terms based on proteins they annotate.

More generally, the function can be used to measure categorical distances between any terms(categories) annotating objects.

Objects should be provided as a first column of a data.table, terms (categories) should be provided as a second column.

Important: for correct evalutation mapping_table should inlude all term-to-gene associations, terms_to_compare should specify which you want to compare
}
\author{
Vitalii Kleshchevnikov
}
