% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler-package.R
\docType{package}
\name{clusterProfiler-package}
\alias{clusterProfiler}
\alias{clusterProfiler-package}
\title{clusterProfiler: A Universal Enrichment Tool for Interpreting Omics Data}
\description{
A universal tool for interpreting functional characteristics of omics data. It supports Over-Representation Analysis (ORA) and Gene Set Enrichment Analysis (GSEA) for both coding and non-coding genomics data of thousands of species. It provides a unified and tidy interface to access, manipulate, and visualize enrichment results. A key capability is the simultaneous analysis and comparison of datasets from multiple treatments or time points. It also provides high-level access to topology-aware enrichment workflows powered by 'enrichit', including single-network and multi-layer network enrichment for Gene Ontology and KEGG analyses. Furthermore, it integrates Large Language Model (LLM) capabilities to provide automated and insightful interpretation of enrichment results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/contribution-knowledge-mining/}
  \item Report bugs at \url{https://github.com/YuLab-SMU/clusterProfiler/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Li-Gen Wang \email{reeganwang020@gmail.com} [contributor]
  \item Xiao Luo \email{l77880853349@163.com} [contributor]
  \item Meijun Chen \email{mjchen1996@outlook.com} [contributor]
  \item Giovanni Dall'Olio \email{giovanni.dallolio@upf.edu} [contributor]
  \item Wanqian Wei \email{altair_wei@outlook.com} [contributor]
  \item Chun-Hui Gao \email{gaospecial@gmail.com} (\href{https://orcid.org/0000-0002-1445-7939}{ORCID}) [contributor]
}

}
\keyword{internal}
