% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cnetplot}
\alias{cnetplot}
\title{cnetplot}
\usage{
cnetplot(x, ...)
}
\arguments{
\item{x}{enrichResult object}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
category-gene-net plot
}
\details{
category gene association
}
\author{
ygc
}

