\name{compareClusterResult-class}
\title{Class "compareClusterResult"...}
\description{Class "compareClusterResult"
This class represents the comparison result of gene clusters by GO
categories at specific level or GO enrichment analysis.}
\alias{compareClusterResult-class}
\alias{show,compareClusterResult-method}
\alias{summary,compareClusterResult-method}
\alias{plot,compareClusterResult-method}
\docType{class}
\author{Guangchuang Yu \url{http://ygc.name}}
\seealso{\code{\linkS4class{groupGOResult}}
\code{\linkS4class{enrichGOResult}} \code{\link{compareCluster}}}
\keyword{classes}
