% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCluster.R
\docType{methods}
\name{dotplot,compareClusterResult-method}
\alias{dotplot,compareClusterResult,ANY-method}
\alias{dotplot,compareClusterResult-method}
\title{dotplot}
\usage{
\S4method{dotplot}{compareClusterResult}(object, x = ~Cluster,
  colorBy = "p.adjust", showCategory = 5, by = "geneRatio",
  split = NULL, includeAll = TRUE, font.size = 12, title = "")
}
\arguments{
\item{object}{compareClusterResult object}

\item{x}{x variable}

\item{colorBy}{one of pvalue or p.adjust}

\item{showCategory}{category numbers}

\item{by}{one of geneRatio, Percentage or count}

\item{split}{ONTOLOGY or NULL}

\item{includeAll}{logical}

\item{font.size}{font size}

\item{title}{figure title}
}
\description{
dot plot method
}

