% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DOSE.R
\name{dotplot}
\alias{dotplot}
\title{dotplot}
\usage{
dotplot(object, x = "geneRatio", colorBy = "p.adjust", showCategory = 10,
  font.size = 12, title = "")
}
\arguments{
\item{object}{an instance of enrichResult}

\item{x}{variable for x axis}

\item{colorBy}{one of 'pvalue', 'p.adjust' and 'qvalue'}

\item{showCategory}{number of category}

\item{font.size}{font size}

\item{title}{plot title}
}
\value{
ggplot object
}
\description{
dot plot of enrichResult
}
\author{
Guangchuang Yu
}

