% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGG.R
\name{download.KEGG}
\alias{download.KEGG}
\title{download.KEGG}
\usage{
download.KEGG(species, KEGG_Type = "KEGG", keyType = "kegg")
}
\arguments{
\item{species}{species}

\item{KEGG_Type}{one of 'KEGG' or 'MKEGG'}

\item{keyType}{supported keyType, see bitr_kegg}
}
\value{
list
}
\description{
download the latest version of KEGG pathway
}
\author{
Guangchuang Yu
}

