% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGG.R
\name{download_KEGG}
\alias{download_KEGG}
\title{download_KEGG}
\usage{
download_KEGG(species, keggType = "KEGG", keyType = "kegg")
}
\arguments{
\item{species}{species}

\item{keggType}{one of 'KEGG' or 'MKEGG'}

\item{keyType}{supported keyType, see bitr_kegg}
}
\value{
list
}
\description{
download the latest version of KEGG pathway/module
}
\author{
Guangchuang Yu
}
