% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichGO.R
\name{dropGO}
\alias{dropGO}
\title{dropGO}
\usage{
dropGO(x, level = NULL, term = NULL)
}
\arguments{
\item{x}{an instance of 'enrichResult' or 'compareClusterResult'}

\item{level}{GO level}

\item{term}{GO term}
}
\value{
modified version of x
}
\description{
drop GO term of specific level or specific terms (mostly too general).
}
\author{
Guangchuang Yu
}
