% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichDAVID.R
\name{enrichDAVID}
\alias{enrichDAVID}
\title{enrichDAVID}
\usage{
enrichDAVID(gene, idType = "ENTREZ_GENE_ID", listType = "Gene",
  minGSSize = 5, annotation = "GOTERM_BP_ALL", pvalueCutoff = 0.05,
  pAdjustMethod = "BH", qvalueCutoff = 0.2, species = NA,
  david.user = "clusterProfiler@hku.hk")
}
\arguments{
\item{gene}{input gene}

\item{idType}{id type}

\item{listType}{list Type}

\item{minGSSize}{minGSSize}

\item{annotation}{david annotation}

\item{pvalueCutoff}{pvalueCutoff}

\item{pAdjustMethod}{one of "BH" and "bonferroni"}

\item{qvalueCutoff}{qvalutCutoff}

\item{species}{species}

\item{david.user}{david user}
}
\value{
A \code{enrichResult} instance
}
\description{
enrichment analysis by DAVID
}
\author{
Guangchuang Yu
}

