\name{enrichGOResult-class}
\title{Class "enrichGOResult"...}
\description{Class "enrichGOResult"
This class represents the result of GO enrichment analysis with FDR control.}
\alias{enrichGOResult-class}
\alias{show,enrichGOResult-method}
\alias{summary,enrichGOResult-method}
\alias{plot,enrichGOResult-method}
\docType{class}
\author{Guangchuang Yu \url{http://ygc.name}}
\seealso{\code{\linkS4class{compareClusterResult}}
\code{\link{compareCluster}} \code{\link{enrichGO}}}
\keyword{classes}
