\docType{class}
\name{enrichKEGGResult-class}
\alias{enrichKEGGResult-class}
\alias{plot,enrichKEGGResult-method}
\alias{show,enrichKEGGResult-method}
\alias{summary,enrichKEGGResult-method}
\title{Class "enrichKEGGResult"
This class represents the result of KEGG enrichment analysis.}
\description{
  Class "enrichKEGGResult" This class represents the result
  of KEGG enrichment analysis.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\linkS4class{compareClusterResult}}
  \code{\link{compareCluster}} \code{\link{enrichKEGG}}
}
\keyword{classes}

