% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmbKEGG.R
\name{enrichmbKEGG}
\alias{enrichmbKEGG}
\title{KEGG enrichment analysis for microbiome.}
\usage{
enrichmbKEGG(
  gene,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  use_internal_data = FALSE,
  microbe_data
)
}
\arguments{
\item{gene}{a vector of K gene id (e.g. K00001).}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{universe}{universe background genes. If missing, the all genes listed in the microbe_data will be used as background.}

\item{minGSSize}{minGSSize minimal size of genes annotated by KEGG term for testing.}

\item{maxGSSize}{maxGSSize maximal size of genes annotated for testing.}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report as significant.}

\item{microbe_data}{a list of all K gene numbers of your microbiome data as universe.}
}
\value{
A \code{enrichResult} instance.
}
\description{
KEGG enrichment analysis for microbiome.
}
\examples{
\dontrun{
data(hgmlist)
yy <- enrichmbKEGG(hgmlist[1:20],microbe_data=hgmlist)
}


}
