\name{geneID2geneName}
\alias{geneID2geneName}
\title{convert gene IDs to gene Names}
\usage{
  geneID2geneName(geneID.list, organism)
}
\arguments{
  \item{geneID.list}{a list of gene IDs}

  \item{organism}{one of human, mouse and yeast.}
}
\value{
  a list of gene names.
}
\description{
  convert a list of gene IDs to gene Names.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

