% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{get_ppi_network}
\alias{get_ppi_network}
\title{get_ppi_network}
\usage{
get_ppi_network(
  taxID = 9606,
  version = "12.0",
  score_threshold = 400,
  output = "data.frame"
)
}
\arguments{
\item{taxID}{NCBI taxon identifier (e.g., 9606 for Human).}

\item{version}{STRING database version (default: "12.0").}

\item{score_threshold}{threshold of significance to include a interaction (default: 400).}

\item{output}{one of `data.frame` or `igraph`}
}
\value{
a `data.frame` or an `igraph` object containing the background PPI network.
}
\description{
get_ppi_network
}
\details{
Download and cache the full STRING PPI background network for a specific species.
}
