% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiPathways.R
\name{get_wp_organisms}
\alias{get_wp_organisms}
\title{get_wp_organism}
\usage{
get_wp_organisms()
}
\value{
supported organism list
}
\description{
list supported organism of WikiPathways
}
\details{
This function extracts information from 'https://data.wikipathways.org/current/gmt/'
and lists all supported organisms
}
\author{
Guangchuang Yu
}
