% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{getGeneSet.BP}
\alias{getGeneSet.BP}
\title{getGeneSet.BP}
\usage{
\method{getGeneSet}{BP}(setType = "BP", organism, use_internal_data = TRUE)
}
\arguments{
\item{setType}{gene set type}

\item{organism}{organism}

\item{use_internal_data}{logical}
}
\description{
getGeneSet.BP
}

