% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{getGeneSet.KEGG}
\alias{getGeneSet.KEGG}
\title{getGeneSet.KEGG}
\usage{
\method{getGeneSet}{KEGG}(setType = "KEGG", organism)
}
\arguments{
\item{setType}{gene set type}

\item{organism}{organism}
}
\description{
getGeneSet.KEGG
}

