% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getGOLevel}
\alias{getGOLevel}
\title{get GOIDs at a specific level}
\usage{
getGOLevel(ont, level)
}
\arguments{
\item{ont}{Ontology}

\item{level}{GO level}
}
\value{
a vector of GOIDs
}
\description{
query GOIDs at a specific level.
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

