% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmbKEGG.R
\name{getM_DATA}
\alias{getM_DATA}
\title{get universe microbiome data}
\usage{
getM_DATA(Kgenelist)
}
\arguments{
\item{Kgenelist}{a list of K gene numbers of your microbiome data as universe}
}
\value{
A gene annotation table of  human gut microbiome
}
\description{
get universe microbiome data
}
\examples{
\dontrun{
getM_DATA(hgmlist)
}


}
