\name{getRatio}
\alias{getRatio}
\title{getRatio}
\usage{
  getRatio(x)
}
\arguments{
  \item{x}{vector of numerator and denominator}
}
\value{
  numerator/denominator
}
\description{
  provide numerator and denominator, return
  numerator/denominator
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

