% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{getTaxID}
\alias{getTaxID}
\title{getTaxID}
\usage{
getTaxID(species)
}
\arguments{
\item{species}{scientific name of a species}
}
\value{
taxonomic ID
}
\description{
Convert species scientific name to taxonomic ID
}
\author{
Guangchuang Yu
}
