\name{Gff2GeneTable}
\alias{Gff2GeneTable}
\title{Gff2GeneTable}
\usage{
Gff2GeneTable(gffFile, compress = TRUE)
}
\arguments{
  \item{gffFile}{GFF file}

  \item{compress}{compress file or not}
}
\value{
file save.
}
\description{
read GFF file and build gene information table
}
\details{
given the GFF file, this function will extract information
and save it in working directory
}
\author{
Yu Guangchuang
}

