% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-utilities.R
\name{go2ont}
\alias{go2ont}
\title{go2ont}
\usage{
go2ont(goid)
}
\arguments{
\item{goid}{a vector of GO IDs}
}
\value{
data.frame
}
\description{
convert goid to ontology (BP, CC, MF)
}
\author{
Guangchuang Yu
}
