\name{GO2Term}
\alias{GO2Term}
\title{Mapping GOIDs to GO Terms}
\usage{
  GO2Term(GOID)
}
\arguments{
  \item{GOID}{GOID}
}
\value{
  GO Terms
}
\description{
  provide a vector of GOIDs, this function will convert
  them to corresponding GO Terms
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

