% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofilter.R
\name{gofilter}
\alias{gofilter}
\title{gofilter}
\usage{
gofilter(x, level = 4)
}
\arguments{
\item{x}{output from enrichGO or compareCluster}

\item{level}{GO level}
}
\value{
updated object
}
\description{
filter GO enriched result at specific level
}
\author{
Guangchuang Yu
}
