\docType{class}
\name{groupGOResult-class}
\alias{groupGOResult-class}
\alias{show,groupGOResult-method}
\title{Class "groupGOResult"
This class represents the result of functional Profiles of a set of gene at
specific GO level.}
\description{
  Class "groupGOResult" This class represents the result of
  functional Profiles of a set of gene at specific GO
  level.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\linkS4class{compareClusterResult}}
  \code{\link{compareCluster}} \code{\link{groupGO}}
}
\keyword{classes}

