% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gseaplot}
\alias{gseaplot}
\title{gseaplot}
\usage{
gseaplot(x, ...)
}
\arguments{
\item{x}{gseaResult object}

\item{...}{additional parameters}
}
\value{
figure
}
\description{
visualize analyzing result of GSEA
}
\details{
plotting function for gseaResult
}
\author{
ygc
}

