% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseMeSH}
\alias{gseMeSH}
\title{gseMeSH}
\usage{
gseMeSH(geneList, MeSHDb, database = "gendoo", category = "C",
  exponent = 1, nPerm = 1000, minGSSize = 10, maxGSSize = 500,
  pvalueCutoff = 0.05, pAdjustMethod = "BH", verbose = TRUE,
  seed = FALSE, by = "fgsea")
}
\arguments{
\item{geneList}{order ranked geneList}

\item{MeSHDb}{MeSHDb}

\item{database}{one of 'gendoo', 'gene2pubmed' or 'RBBH'}

\item{category}{one of "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L","M", "N", "V", "Z"}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{seed}{logical}

\item{by}{one of 'fgsea' or 'DOSE'}
}
\value{
gseaResult object
}
\description{
Gene Set Enrichment Analysis of MeSH
}
\author{
Yu Guangchuang
}

