% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayCommons.R
\name{gsePC}
\alias{gsePC}
\title{gsePC}
\usage{
gsePC(geneList, eps = 1e-10, ...)
}
\arguments{
\item{geneList}{a ranked gene list}

\item{eps}{boundary for calculating the p value in multilevel mode}

\item{...}{additional parameters, see also the parameters supported by the GSEA() function}
}
\value{
A \code{gseaResult} instance
}
\description{
GSEA analysis for  Pathway Commons
}
\details{
This function performs GSEA using  Pathway Commons
}
