% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson.R
\name{gson_GO}
\alias{gson_GO}
\title{gson_KEGG}
\usage{
gson_GO(OrgDb, keytype = "ENTREZID", ont = "BP")
}
\arguments{
\item{OrgDb}{OrgDb}

\item{keytype}{keytype of genes.}

\item{ont}{one of "BP", "MF", "CC", and "ALL"}
}
\value{
a 'GSON' object
}
\description{
download the latest version of KEGG pathway and stored in a 'GSON' object
}
