% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson.R
\name{gson_GO_local}
\alias{gson_GO_local}
\title{Build a gson object that annotate Gene Ontology}
\usage{
gson_GO_local(data, ont = c("ALL", "BP", "CC", "MF"), species = NULL, ...)
}
\arguments{
\item{data}{a two-column data.frame of original GO annotation. The columns are "gene_id" and "go_id".}

\item{ont}{type of GO annotation, which is "ALL", "BP", "MF", or "CC". default: "ALL".}

\item{species}{name of species. Default: NULL.}

\item{...}{pass to `gson::gson()` constructor.}
}
\value{
a `gson` instance
}
\description{
Build a gson object that annotate Gene Ontology
}
\examples{
 data = data.frame(gene_id = "gene1",
                   go_id = c("GO:0035492", "GO:0009764", "GO:0031063", "GO:0033714", "GO:0036349"))
 gson_GO_local(data, species = "E. coli")
}
