% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson.R
\name{gson_WP}
\alias{gson_WP}
\title{gson_WP}
\usage{
gson_WP(organism)
}
\arguments{
\item{organism}{supported organism, which can be accessed via the get_wp_organisms() function.}
}
\description{
Download the latest version of WikiPathways data and stored in a 'GSON' object
}
