% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgmlist}
\alias{hgmlist}
\title{A list with 6980 K genes of human gut genome}
\format{
A list with 6980 K genes
}
\usage{
hgmlist
}
\description{
All 6980 K genes were reported on
"An integrated catalog of reference genes in the
 human gut microbiome.[J]. Nature Biotechnology, 2014."
}
\details{
data(hgmlist)
 kk <- enrichKEGG(gene = hgmlist[1:20], organism = "ko", universe = hgmlist )
 dotplot(kk)
}
\keyword{datasets}
