\name{HyperG}
\alias{HyperG}
\title{hypergeometric test}
\usage{HyperG(numWdrawn, numW, numB, numDrawn)}
\description{hypergeometric test for enrichment analysis}
\value{pvalue}
\author{Guangchuang Yu \url{http://ygc.name}}
\arguments{\item{numWdrawn}{number of White balls drawn}
\item{numW}{number of White balls}
\item{numB}{number of Black balls}
\item{numDrawn}{number of balls drawn}}
