% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitr.R
\name{idType}
\alias{idType}
\title{idType}
\usage{
idType(annoDb = "org.Hs.eg.db")
}
\arguments{
\item{annoDb}{annotation db}
}
\value{
character vector
}
\description{
list ID types supported by annoDb
}
\author{
Guangchuang Yu
}

