% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret_hierarchical}
\alias{interpret_hierarchical}
\title{interpret_hierarchical}
\usage{
interpret_hierarchical(
  x_minor,
  x_major,
  mapping,
  model = "deepseek-chat",
  provider = NULL,
  api_key = NULL,
  task = "cell_type"
)
}
\arguments{
\item{x_minor}{Enrichment result for sub-clusters (e.g., compareClusterResult or list of enrichResult).}

\item{x_major}{Enrichment result for major clusters.}

\item{mapping}{A named vector mapping sub-cluster IDs (names in x_minor) to major cluster IDs (names in x_major).}

\item{model}{LLM model.}

\item{api_key}{API key.}

\item{task}{Task type, default is "cell_type".}
}
\value{
A list of interpretation results.
}
\description{
Interpret enrichment results using a hierarchical strategy (Major -> Minor clusters)
}
\author{
Guangchuang Yu
}
