% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_score.R
\name{kappa_score}
\alias{kappa_score}
\title{similarity between categories - kappa score}
\usage{
kappa_score(value)
}
\arguments{
\item{value}{a data.table with 2 columns (two categories) and many rows (objects)}
}
\description{
\code{kappa_score} calculates kappa similarity score between two categories based on the overlap of the set of elements they describe adjusted for that overlap being produced by chance.
}
\details{
Categories can be anything, common example is GO terms (describe gene products, \url{http://geneontology.org/}), KEGG pathways (describe genes, \url{http://www.genome.jp/kegg/pathway.html}), genes (similarity based on annotation with GO terms).

\code{kappa_score} takes a data.table with 2 columns describing to two categories to which each element in rows can belong to. Os and 1s in the table signify membership.

\code{kappa_score} is defined as described in this article: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2375021/figure/F2/}

relies on data.table for subsetting and vcd package for calculating kappa score given 2x2 incidence matrix

\code{kappa_score} is faster than irr::kappa2(ratings = value, weight = "unweighted")

\code{kappa_score} is ~25 times slower than dist()
}
\seealso{
\code{\link{categ_dist}}
}
\author{
Vitalii Kleshchevnikov
}
