% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitr.R
\name{KEGG_path2extid}
\alias{KEGG_path2extid}
\title{KEGG_path2extid}
\usage{
KEGG_path2extid(keggID, species = sub("\\\\d+$", "", keggID),
  keggType = "Path", keyType = "kegg")
}
\arguments{
\item{keggID}{KEGG ID, path or module ID}

\item{species}{species}

\item{keggType}{one of 'Path' or 'Module'}

\item{keyType}{KEGG gene type, one of "ncbi-proteinid", "ncbi-geneid", "uniprot", or "kegg"}
}
\value{
extid vector
}
\description{
query all genes in a KEGG pathway or module
}
\author{
guangchuang yu
}
