\name{list2graph}
\alias{list2graph}
\title{convert gene IDs to igraph object}
\usage{
  list2graph(inputList)
}
\arguments{
  \item{inputList}{a list of gene IDs}
}
\value{
  a igraph object.
}
\description{
  convert a list of gene IDs to igraph object.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

