% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCluster.R
\name{merge_result}
\alias{merge_result}
\title{merge_result}
\usage{
merge_result(enrichResultList)
}
\arguments{
\item{enrichResultList}{a list of enrichResult objects}
}
\value{
a compareClusterResult instance
}
\description{
merge a list of enrichResult objects to compareClusterResult
}
\author{
Guangchuang Yu
}
