% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nseaAnalyzer.R
\name{nseGO}
\alias{nseGO}
\title{nseGO}
\usage{
nseGO(
  geneList,
  network,
  ont = "BP",
  OrgDb,
  keyType = "ENTREZID",
  mode = c("evidence", "signed"),
  p = 0.5,
  specific_weight = FALSE,
  minGSSize = 10,
  maxGSSize = 500,
  threshold = 1e-09,
  maxIter = 100,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{network}{edge list (data.frame with 2 or 3 columns) or sparse matrix}

\item{ont}{one of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{OrgDb}{OrgDb}

\item{keyType}{keytype of gene}

\item{mode}{propagation mode, one of "evidence" and "signed"}

\item{p}{restart probability for RWR}

\item{specific_weight}{logical, whether to apply gene specificity weighting}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{threshold}{convergence threshold for RWR}

\item{maxIter}{maximal number of RWR iterations}

\item{verbose}{print message or not}

\item{...}{other parameters passed to \code{enrichit::nsea_gson()}}
}
\value{
nseaResult object
}
\description{
Network-based Gene Set Enrichment Analysis of Gene Ontology
}
