% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nseaAnalyzer.R
\name{nseKEGG}
\alias{nseKEGG}
\title{nseKEGG}
\usage{
nseKEGG(
  geneList,
  network,
  organism = "hsa",
  keyType = "kegg",
  mode = c("evidence", "signed"),
  p = 0.5,
  specific_weight = FALSE,
  minGSSize = 10,
  maxGSSize = 500,
  threshold = 1e-09,
  maxIter = 100,
  verbose = TRUE,
  use_internal_data = FALSE,
  ...
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{network}{edge list (data.frame with 2 or 3 columns) or sparse matrix}

\item{organism}{supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html',
or a GSON object}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncbi-proteinid' and 'uniprot'}

\item{mode}{propagation mode, one of "evidence" and "signed"}

\item{p}{restart probability for RWR}

\item{specific_weight}{logical, whether to apply gene specificity weighting}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{threshold}{convergence threshold for RWR}

\item{maxIter}{maximal number of RWR iterations}

\item{verbose}{print message or not}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}

\item{...}{other parameters passed to \code{enrichit::nsea_gson()}}
}
\value{
nseaResult object
}
\description{
Network-based Gene Set Enrichment Analysis of KEGG
}
