% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGAF.R
\name{parse.GAF}
\alias{parse.GAF}
\title{parse.GAF}
\usage{
parse.GAF(gafFile, nrows = -1)
}
\arguments{
\item{gafFile}{GAF file}

\item{nrows}{a parameter}
}
\value{
a list with two dataframes
}
\description{
parse GAF file
}
\details{
given a GAF file, this function extracts the information from it
}
