% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_parse_gaf_file.R
\name{parse_gff}
\alias{parse_gff}
\title{parse_gff}
\usage{
parse_gff(GafFile, nrows = -1)
}
\arguments{
\item{GafFile}{GAF file}

\item{nrows}{a parameter}
}
\value{
a list with two dataframes
}
\description{
new parse GAF files
}
\details{
given a GAF file, this function extracts the information from it
}
