% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGAF.R
\name{ParseGafFile}
\alias{ParseGafFile}
\title{ParseGafFile}
\usage{
ParseGafFile(GafFile, nrows = -1)
}
\arguments{
\item{GafFile}{GAF file}

\item{nrows}{a parameter}
}
\value{
a list with two dataframes
}
\description{
parse GAF files
}
\details{
given a GAF file, this function extracts the information from it
}
