\name{path2Name}
\alias{path2Name}
\title{convert KEGG pathway ID to pathway Name}
\usage{
  path2Name(pathIDs)
}
\arguments{
  \item{pathIDs}{KEGG pathway IDs}
}
\value{
  KEGG pathway names
}
\description{
  provide a vector of KEGG pathway IDs, this function will
  convert them to corresponding KEGG pathway Names
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

