\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,compareClusterResult,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{compareClusterResult,ANY}(x, type = "dot", title = "",
  font.size = 12, showCategory = 5, by = "geneRatio",
  colorBy = "p.adjust")
}
\arguments{
  \item{...}{Additional argument list}

  \item{x}{the coordinates of points in the plot.
  Alternatively, a single plotting structure, function or
  \emph{any \R object with a \code{plot} method} can be
  provided.}
}
\value{
plot
}
\description{
plot method generics
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

