% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interpret.R
\name{plot.interpretation}
\alias{plot.interpretation}
\title{plot}
\usage{
\method{plot}{interpretation}(x, layout = "nicely", ...)
}
\arguments{
\item{x}{An `interpretation` object.}

\item{layout}{Graph layout, default is "nicely".}

\item{...}{Additional arguments passed to `ggplot2::ggplot`.}
}
\description{
plot
}
