\name{plot}
\alias{plot}
\title{plot method}
\description{plot method for \code{groupGOResult} instance}
\docType{methods}
\value{ggplot object}
\author{GuangchuangYu}
\arguments{\item{x}{A \code{groupGOResult} instance}
\item{order}{logical parameter, order the result by *Count*.}
\item{title}{graph title}
\item{font.size}{graph font size}
\item{drop}{logical parameter, drop void category.}}
