% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOgraph.R
\name{plotGOgraph}
\alias{plotGOgraph}
\title{plotGOgraph}
\usage{
plotGOgraph(x, firstSigNodes = 10, useInfo = "all", sigForAll = TRUE,
  useFullNames = TRUE, ...)
}
\arguments{
\item{x}{output of enrichGO or gseGO}

\item{firstSigNodes}{number of significant nodes (retangle nodes in the graph)}

\item{useInfo}{additional info}

\item{sigForAll}{if TRUE the score/p-value of all nodes in the DAG is shown, otherwise only score will be shown}

\item{useFullNames}{logical}

\item{...}{additional parameter of showSigOfNodes, please refer to topGO}
}
\value{
GO DAG graph
}
\description{
plot GO graph
}
\seealso{
\link[topGO]{showSigOfNodes}
}
\author{
Guangchuang Yu
}
