% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGAF.R
\name{read.gaf}
\alias{read.gaf}
\alias{parse_gff}
\title{read.gaf}
\usage{
read.gaf(GafFile)

parse_gff(GafFile)
}
\arguments{
\item{GafFile}{GAF file}
}
\value{
a list with two dataframes
}
\description{
parse GAF files
}
\details{
given a GAF file, this function extracts the information from it and add indirect GO annotation
}
