% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{read.gmt}
\alias{read.gmt}
\alias{read.gmt.wp}
\title{read.gmt}
\usage{
read.gmt(gmtfile)

read.gmt.wp(gmtfile)
}
\arguments{
\item{gmtfile}{gmt file}
}
\value{
data.frame
}
\description{
parse gmt file to a data.frame
}
\author{
Guangchuang Yu
}
