% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayCommons.R
\name{read.gmt.pc}
\alias{read.gmt.pc}
\title{read.gmt.pc}
\usage{
read.gmt.pc(gmtfile, output = "data.frame")
}
\arguments{
\item{gmtfile}{A gmt file}

\item{output}{one of 'data.frame' or 'GSON'}
}
\value{
A data.frame or A GSON object depends on the value of 'output'
}
\description{
Parse gmt file from Pathway Common
}
\details{
This function parse gmt file downloaded from Pathway common
}
