% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg-utilities.R
\name{search_kegg_organism}
\alias{search_kegg_organism}
\title{search_kegg_organism}
\usage{
search_kegg_organism(str, by = "scientific_name", ignore.case = FALSE)
}
\arguments{
\item{str}{string}

\item{by}{one of 'kegg.code', 'scientific_name' and 'common_name'}

\item{ignore.case}{TRUE or FALSE}
}
\value{
data.frame
}
\description{
search kegg organism, listed in http://www.genome.jp/kegg/catalog/org_list.html
}
\author{
Guangchuang Yu
}
