\docType{methods}
\name{show}
\alias{show}
\title{show method}
\arguments{
  \item{object}{A \code{compareClusterResult} instance.}

  \item{object}{A \code{groupGOResult} instance}
}
\value{
  message

  message
}
\description{
  show method for \code{compareClusterResult} instance

  show method for \code{groupGOResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}

  Guangchuang Yu \url{http://ygc.name}
}

