% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\docType{methods}
\name{simplify}
\alias{simplify}
\alias{simplify,enrichResult-method}
\alias{simplify}
\alias{simplify,gseaResult-method}
\alias{simplify}
\alias{simplify,compareClusterResult-method}
\title{simplify method}
\usage{
\S4method{simplify}{enrichResult}(x, cutoff = 0.7, by = "p.adjust",
  select_fun = min, measure = "Wang", semData = NULL,
  use_data_table = ifelse(measure == "kappa", T, F),
  use_bioc_annotationdbi = T)

\S4method{simplify}{gseaResult}(x, cutoff = 0.7, by = "p.adjust",
  select_fun = min, measure = "Wang", semData = NULL,
  use_data_table = ifelse(measure == "kappa", T, F),
  use_bioc_annotationdbi = T)

\S4method{simplify}{compareClusterResult}(x, cutoff = 0.7, by = "p.adjust",
  select_fun = min, measure = "Wang", semData = NULL,
  use_data_table = ifelse(measure == "kappa", T, F),
  use_bioc_annotationdbi = T)
}
\arguments{
\item{x}{output of enrichGO}

\item{cutoff}{similarity cutoff}

\item{by}{feature to select representative term, selected by 'select_fun' function}

\item{select_fun}{function to select feature passed by 'by' parameter}

\item{measure}{method to measure similarity}

\item{semData}{GOSemSimDATA object}

\item{use_data_table}{whether to use data.table for table manipulations}

\item{use_bioc_annotationdbi}{whether to use AnnotationDbi for entity to GO term mapping (currently no alternative)}

\item{x}{output of gseaResult}

\item{cutoff}{similarity cutoff}

\item{by}{feature to select representative term, selected by 'select_fun' function}

\item{select_fun}{function to select feature passed by 'by' parameter}

\item{measure}{method to measure similarity}

\item{semData}{GOSemSimDATA object}

\item{use_data_table}{whether to use data.table for table manipulations}

\item{use_bioc_annotationdbi}{whether to use AnnotationDbi for entity to GO term mapping (currently no alternative)}
}
\value{
updated enrichResult object

updated gseaResult object

updated compareClusterResult object
}
\description{
simplify output from enrichGO by removing redundancy of enriched GO terms

simplify output from gseaResult by removing redundancy of enriched GO terms

simplify output from compareCluster by removing redundancy of enriched GO terms
}
\references{
issue #28
\url{https://github.com/GuangchuangYu/clusterProfiler/issues/28}

issue #28
\url{https://github.com/GuangchuangYu/clusterProfiler/issues/28}
}
\author{
Guangchuang Yu, edits by Vitalii Kleshchevnikov

Guangchuang Yu, adopted for gseaResult by Vitalii Kleshchevnikov

Guangchuang Yu
}
