\docType{methods}
\name{summary}
\alias{summary}
\title{summary method}
\arguments{
  \item{object}{A \code{compareClusterResult} instance.}

  \item{object}{A \code{enrichGOResult} instance.}

  \item{object}{A \code{enrichKEGGResult} instance.}

  \item{object}{A \code{groupGOResult} instance}
}
\value{
  A data frame

  A data frame

  A data frame

  A data frame
}
\description{
  summary method for \code{compareClusterResult} instance

  summary method for \code{enrichGOResult} instance

  summary method for \code{enrichKEGGResult} instance

  summary method for \code{groupGOResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}

  GuangchuangYu \url{http://ygc.name}

  Guangchuang Yu \url{http://ygc.name}

  Guangchuang Yu
}

