% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_get}
\alias{uniprot_get}
\title{uniprot_get}
\usage{
uniprot_get(taxID)
}
\arguments{
\item{taxID}{taxonomy ID}
}
\value{
gene table data frame
}
\description{
retreve annotation data from uniprot
}
\author{
guangchuang yu
}
