% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitr.R
\name{bitr_kegg}
\alias{bitr_kegg}
\title{bitr_kegg}
\usage{
bitr_kegg(geneID, fromType, toType, organism, drop = TRUE)
}
\arguments{
\item{geneID}{input gene id}

\item{fromType}{input id type}

\item{toType}{output id type}

\item{organism}{supported organism, can be search using search_kegg_organism function}

\item{drop}{drop NA or not}
}
\value{
data.frame
}
\description{
convert biological ID using KEGG API
}
\author{
Guangchuang Yu
}
