% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{buildGOmap}
\alias{buildGOmap}
\title{buildGOmap}
\usage{
buildGOmap(gomap, compress = TRUE)
}
\arguments{
\item{gomap}{data.frame with two columns names "entrezgene", and "go_accession"}

\item{compress}{logical, indicate file save in compress or not.}
}
\value{
files save in the the working directory
}
\description{
building GO mapping files
}
\details{
provided by a data.frame of gene and GO directly annotation file
this function will building gene to GO and GO to gene mapping,
with directly and undirectly annotation.
}
\author{
Yu Guangchuang
}

