% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOmap.R
\name{buildGOmap}
\alias{buildGOmap}
\title{buildGOmap}
\usage{
buildGOmap(gomap)
}
\arguments{
\item{gomap}{data.frame with two columns of GO and gene ID}
}
\value{
data.frame, GO annotation with indirect annotation
}
\description{
building GO mapping files
}
\details{
provided by a data.frame of GO (column 1) and gene (column 2) direct annotation
this function will building gene to GO and GO to gene mapping,
with directly and undirectly (ancestor GO term) annotation.
}
\author{
Yu Guangchuang
}
