% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{buildKEGGmap}
\alias{buildKEGGmap}
\title{buildKEGGmap}
\usage{
buildKEGGmap(keggmap, id2name = NULL, organism)
}
\arguments{
\item{keggmap}{pathway to external ID}

\item{id2name}{pathway id to pathway name}

\item{organism}{organism}
}
\description{
build KEGG annotation files
}
\author{
Guangchuang Yu
}

