% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler-package.R
\docType{package}
\name{clusterProfiler-package}
\alias{clusterProfiler-package}
\alias{clusterProfiler}
\title{statistical analysis and visualization of functional profiles for genes and
gene clusters}
\description{
The package implements methods to analyze and visualize functional profiles
of gene and gene clusters.
}
