\docType{package}
\name{clusterProfiler-package}
\alias{clusterProfiler}
\alias{clusterProfiler-package}
\title{statistical analysis and visulization of functional profiles for genes and
gene clusters
The package implements methods to analyze and visualize functional profiles
of gene and gene clusters.}
\description{
  This package is designed to compare gene clusters
  functional profiles.
}
\details{
  \tabular{ll}{ Package: \tab clusterProfiler\cr Type: \tab
  Package\cr Version: \tab 1.9.\cr Date: \tab 06-13-2013\cr
  biocViews:\tab GO, Clustering, Visulization\cr
  Depends:\tab AnnotationDbi, GO.db, org.Hs.eg.db, ggplot2,
  plyr, methods\cr Suggests:\tab GOSemSim\cr License: \tab
  Artistic-2.0\cr }
}
\author{
  Guangchuang Yu \email{guangchuangyu@gmail.com}

  Maintainer: Guangchuang Yu
  \email{guangchuangyu@gmail.com}
}
\seealso{
  \linkS4class{compareClusterResult},
  \linkS4class{groupGOResult} \linkS4class{enrichResult}
}
\keyword{package}

