% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichGO.R
\name{cnetplot}
\alias{cnetplot}
\title{cnetplot}
\usage{
cnetplot(x, ...)
}
\arguments{
\item{x}{enrichResult object}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
category-gene-net plot
}
\details{
category gene association
}
\author{
ygc
}

