% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSE.R
\name{cnetplot}
\alias{cnetplot}
\title{cnetplot}
\usage{
cnetplot(x, showCategory = 5, categorySize = "geneNum", foldChange = NULL,
  fixed = TRUE, ...)
}
\arguments{
\item{x}{enrichResult object}

\item{showCategory}{number of category plotted}

\item{categorySize}{one of geneNum or pvalue}

\item{foldChange}{fold change of expression value}

\item{fixed}{logical}

\item{...}{additional parameter}
}
\value{
plot
}
\description{
category-gene-net plot
}
\details{
category gene association
}
\author{
ygc
}

