% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{compareClusterResult-class}
\alias{compareClusterResult-class}
\alias{show,compareClusterResult-method}
\alias{summary,compareClusterResult-method}
\alias{plot,compareClusterResult-method}
\title{Class "compareClusterResult"
This class represents the comparison result of gene clusters by GO
categories at specific level or GO enrichment analysis.}
\description{
Class "compareClusterResult"
This class represents the comparison result of gene clusters by GO
categories at specific level or GO enrichment analysis.
}
\section{Slots}{

\describe{
\item{\code{compareClusterResult}}{cluster comparing result}

\item{\code{geneClusters}}{a list of genes}

\item{\code{fun}}{one of groupGO, enrichGO and enrichKEGG}

\item{\code{.call}}{function call}
}}

\seealso{
\code{\linkS4class{groupGOResult}}
  \code{\linkS4class{enrichResult}} \code{\link{compareCluster}}
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
\keyword{classes}
