% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler-package.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{gcSample}
\alias{kegg_species}
\title{Datasets
gcSample contains a sample of gene clusters.}
\description{
Datasets
gcSample contains a sample of gene clusters.

Datasets
kegg_species contains kegg species information
}
\keyword{datasets}
