% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCluster.R
\docType{methods}
\name{dotplot,compareClusterResult-method}
\alias{dotplot,compareClusterResult-method}
\alias{dotplot,compareClusterResult,ANY-method}
\title{dotplot}
\usage{
\S4method{dotplot}{compareClusterResult}(object, x = ~Cluster,
  color = "p.adjust", showCategory = 5, split = NULL,
  font.size = 12, title = "", by = "geneRatio", includeAll = TRUE)
}
\arguments{
\item{object}{compareClusterResult object}

\item{x}{x variable}

\item{color}{one of pvalue or p.adjust}

\item{showCategory}{category numbers}

\item{split}{ONTOLOGY or NULL}

\item{font.size}{font size}

\item{title}{figure title}

\item{by}{one of geneRatio, Percentage or count}

\item{includeAll}{logical}
}
\description{
dot plot method
}
