% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{download.KEGG}
\alias{download.KEGG}
\title{download.KEGG}
\usage{
download.KEGG(species)
}
\arguments{
\item{species}{species}
}
\value{
list
}
\description{
download the latest version of KEGG pathway
}
\author{
Guangchuang Yu
}

