% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichDAVID.R
\name{enrichDAVID}
\alias{enrichDAVID}
\title{enrichDAVID}
\usage{
enrichDAVID(gene, idType = "ENTREZ_GENE_ID", universe, minGSSize = 10,
  maxGSSize = 500, annotation = "GOTERM_BP_FAT", pvalueCutoff = 0.05,
  pAdjustMethod = "BH", qvalueCutoff = 0.2, species = NA, david.user)
}
\arguments{
\item{gene}{input gene}

\item{idType}{id type}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated for testing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{annotation}{david annotation}

\item{pvalueCutoff}{pvalueCutoff}

\item{pAdjustMethod}{one of "BH" and "bonferroni"}

\item{qvalueCutoff}{qvalutCutoff}

\item{species}{species}

\item{david.user}{david user}
}
\value{
A \code{enrichResult} instance
}
\description{
enrichment analysis by DAVID
}
\author{
Guangchuang Yu
}
