% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichGO.R
\name{enrichGO}
\alias{enrichGO}
\title{GO Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enrichment GO
categories after FDR control.}
\usage{
enrichGO(gene, OrgDb, keytype = "ENTREZID", ont = "MF",
  pvalueCutoff = 0.05, pAdjustMethod = "BH", universe, qvalueCutoff = 0.2,
  minGSSize = 10, maxGSSize = 500, readable = FALSE)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{OrgDb}{OrgDb}

\item{keytype}{keytype of input gene}

\item{ont}{One of "MF", "BP", and "CC" subontologies.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{qvalueCutoff}{qvalue cutoff}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{readable}{whether mapping gene ID to gene Name}
}
\value{
An \code{enrichResult} instance.
}
\description{
GO Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enrichment GO
categories after FDR control.
}
\examples{
\dontrun{
	data(gcSample)
	yy <- enrichGO(gcSample[[1]], 'org.Hs.eg.db', ont="BP", pvalueCutoff=0.01)
	head(summary(yy))
	plot(yy)
}
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{enrichResult-class}}, \code{\link{compareCluster}}
}
\keyword{manip}

