\docType{class}
\name{enrichGOResult-class}
\alias{enrichGOResult-class}
\alias{plot,enrichGOResult-method}
\alias{show,enrichGOResult-method}
\alias{summary,enrichGOResult-method}
\title{Class "enrichGOResult"
This class represents the result of GO enrichment analysis with FDR control.}
\description{
  Class "enrichGOResult" This class represents the result
  of GO enrichment analysis with FDR control.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\linkS4class{compareClusterResult}}
  \code{\link{compareCluster}} \code{\link{enrichGO}}
}
\keyword{classes}

