% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGG.R
\name{enrichKEGG}
\alias{enrichKEGG}
\title{KEGG Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enrichment KEGG
categories with FDR control.}
\usage{
enrichKEGG(gene, organism = "human", pvalueCutoff = 0.05,
  pAdjustMethod = "BH", universe, minGSSize = 5, qvalueCutoff = 0.2,
  readable = FALSE, use_internal_data = FALSE)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{organism}{One of "anopheles", "arabidopsis", "bovine", "canine",
"chicken", "chimp", "ecolik12","ecsakai", "fly", "human",
"malaria", "mouse", "pig", "rat","rhesus", "worm", "xenopus",
"yeast" and "zebrafish".}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{qvalueCutoff}{qvalue cutoff}

\item{readable}{whether mapping gene ID to gene Name}

\item{use_internal_data}{logical, if TRUE, use KEGG.db.
default is FALSE, will download online KEGG data}
}
\value{
A \code{enrichResult} instance.
}
\description{
KEGG Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enrichment KEGG
categories with FDR control.
}
\examples{

	data(gcSample)
	yy = enrichKEGG(gcSample[[5]], pvalueCutoff=0.01)
	head(summary(yy))
	#plot(yy)

}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{enrichResult-class}}, \code{\link{compareCluster}}
}
\keyword{manip}

