\name{enrichKEGGResult-class}
\title{Class "enrichKEGGResult"...}
\description{Class "enrichKEGGResult"
This class represents the result of KEGG enrichment analysis.}
\alias{enrichKEGGResult-class}
\alias{show,enrichKEGGResult-method}
\alias{summary,enrichKEGGResult-method}
\alias{plot,enrichKEGGResult-method}
\docType{class}
\author{Guangchuang Yu <guangchuangyu@gmail.com>}
\seealso{\code{\linkS4class{compareClusterResult}}
\code{\link{compareCluster}} \code{\link{enrichKEGG}}}
\keyword{classes}
