% Generated by roxygen2 (4.0.1): do not edit by hand
\name{enrichMap}
\alias{enrichMap}
\title{enrichMap}
\usage{
enrichMap(x, ...)
}
\arguments{
\item{x}{enrichResult or gseaResult}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
enrichment map
}
\details{
enrichMap
}
\author{
ygc
}

