% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSE.R
\name{enrichMap}
\alias{enrichMap}
\title{enrichMap}
\usage{
enrichMap(x, n = 50, fixed = TRUE, vertex.label.font = 1, ...)
}
\arguments{
\item{x}{gseaResult or enrichResult object}

\item{n}{maximum number of category to shown}

\item{fixed}{if set to FALSE, will invoke tkplot}

\item{vertex.label.font}{font size of vertex label}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
enrichment map
}
\details{
enrichMap
}
\author{
ygc
}

