% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichGO.R
\name{enrichMap}
\alias{enrichMap}
\title{enrichMap}
\usage{
enrichMap(x, ...)
}
\arguments{
\item{x}{enrichResult or gseaResult}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
enrichment map
}
\details{
enrichMap
}
\author{
ygc
}

