% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMKEGG.R
\name{enrichMKEGG}
\alias{enrichMKEGG}
\title{KEGG Module Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enrichment KEGG Module
categories with FDR control.}
\usage{
enrichMKEGG(gene, organism = "hsa", keyType = "kegg",
  pvalueCutoff = 0.05, pAdjustMethod = "BH", universe,
  minGSSize = 10, maxGSSize = 500, qvalueCutoff = 0.2)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{qvalueCutoff}{qvalue cutoff}
}
\value{
A \code{enrichResult} instance.
}
\description{
KEGG Module Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enrichment KEGG Module
categories with FDR control.
}
