% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayCommons.R
\name{enrichPC}
\alias{enrichPC}
\title{enrichPC}
\usage{
enrichPC(gene, source, keyType = "hgnc", ...)
}
\arguments{
\item{gene}{a vector of entrez gene id}

\item{...}{additional parameters, see also the parameters supported by the enricher() function}

\item{organism}{supported organisms, which can be accessed via the get_pc_organisms() function}
}
\value{
A \code{enrichResult} instance
}
\description{
ORA analysis for Pathway Commons
}
\details{
This function performs over-representation analysis using  Pathway Commons
}
