% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiPathways.R
\name{enrichWP}
\alias{enrichWP}
\title{enrichWP}
\usage{
enrichWP(gene, organism, ...)
}
\arguments{
\item{gene}{a vector of entrez gene id}

\item{organism}{supported organisms, which can be accessed via the get_wp_organisms() function}

\item{...}{additional parameters, see also the parameters supported by the enricher() function}
}
\value{
A \code{enrichResult} instance
}
\description{
ORA analysis for WikiPathways
}
\details{
This function performs over-representation analysis using WikiPathways
}
\author{
Guangchuang Yu
}
