% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{getGeneSet.CC}
\alias{getGeneSet.CC}
\title{getGeneSet.CC}
\usage{
\method{getGeneSet}{CC}(setType = "CC", organism)
}
\arguments{
\item{setType}{gene set type}

\item{organism}{organism}
}
\description{
getGeneSet.CC
}

