% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-utilities.R
\name{getGOLevel}
\alias{getGOLevel}
\title{get GOIDs at a specific level}
\usage{
getGOLevel(ont, level)
}
\arguments{
\item{ont}{Ontology}

\item{level}{GO level}
}
\value{
a vector of GOIDs
}
\description{
query GOIDs at a specific level.
}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
