\name{getRatio}
\alias{getRatio}
\title{getRatio}
\usage{
  getRatio(a, b)
}
\arguments{
  \item{a}{numerator}

  \item{b}{denominator}
}
\value{
  numerator/denominator
}
\description{
  provide numerator and denominator, return
  numerator/denominator
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

