% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFFparser.R
\name{Gff2GeneTable}
\alias{Gff2GeneTable}
\title{Gff2GeneTable}
\usage{
Gff2GeneTable(gffFile, compress = TRUE)
}
\arguments{
\item{gffFile}{GFF file}

\item{compress}{compress file or not}
}
\value{
file save.
}
\description{
read GFF file and build gene information table
}
\details{
given a GFF file, this function extracts information from it and save it in working directory
}
\author{
Yu Guangchuang
}
