% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-utilities.R
\name{go2term}
\alias{go2term}
\title{go2term}
\usage{
go2term(goid)
}
\arguments{
\item{goid}{a vector of GO IDs}
}
\value{
data.frame
}
\description{
convert goid to descriptive term
}
\author{
Guangchuang Yu
}
