\name{groupGOResult-class}
\title{Class "groupGOResult"...}
\description{Class "groupGOResult"
This class represents the result of functional Profiles of a set of gene at
specific GO level.}
\alias{groupGOResult-class}
\alias{show,groupGOResult-method}
\alias{summary,groupGOResult-method}
\alias{plot,groupGOResult-method}
\docType{class}
\author{Guangchuang Yu \url{http://ygc.name}}
\seealso{\code{\linkS4class{compareClusterResult}}
\code{\link{compareCluster}} \code{\link{groupGO}}}
\keyword{classes}
