% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{groupGOResult-class}
\alias{groupGOResult-class}
\alias{show,groupGOResult-method}
\title{Class "groupGOResult"
This class represents the result of functional Profiles of a set of gene at
specific GO level.}
\description{
Class "groupGOResult"
This class represents the result of functional Profiles of a set of gene at
specific GO level.
}
\section{Slots}{

\describe{
\item{\code{result}}{GO classification result}

\item{\code{ontology}}{Ontology}

\item{\code{level}}{GO level}

\item{\code{organism}}{one of "human", "mouse" and "yeast"}

\item{\code{gene}}{Gene IDs}

\item{\code{readable}}{logical flag of gene ID in symbol or not.}
}}

\seealso{
\code{\linkS4class{compareClusterResult}}
  \code{\link{compareCluster}} \code{\link{groupGO}}
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
\keyword{classes}
