% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enricher.R
\name{GSEA}
\alias{GSEA}
\title{GSEA}
\usage{
GSEA(geneList, exponent = 1, nPerm = 1000, minGSSize = 10,
  pvalueCutoff = 0.05, pAdjustMethod = "BH", TERM2GENE, TERM2NAME = NA,
  verbose = TRUE, seed = FALSE)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{exponent}{weight of each step}

\item{nPerm}{number of permutations}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{pvalueCutoff}{pvalue cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{TERM2GENE}{user input annotation of TERM TO GENE mapping, a data.frame of 2 column with term and gene}

\item{TERM2NAME}{user input of TERM TO NAME mapping, a data.frame of 2 column with term and name}

\item{verbose}{logical}

\item{seed}{logical}
}
\value{
gseaResult object
}
\description{
a universal gene set enrichment analysis tools
}
\author{
Guangchuang Yu
}

