% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseaplot}
\alias{gseaplot}
\title{gseaplot}
\usage{
gseaplot(gseaResult, geneSetID, by = "all")
}
\arguments{
\item{gseaResult}{gseaResult object}

\item{geneSetID}{geneSet ID}

\item{by}{one of "runningScore" or "position"}
}
\value{
ggplot2 object
}
\description{
visualize analyzing result of GSEA
}
\details{
plotting function for gseaResult
}
\author{
ygc
}

