% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseGO}
\alias{gseGO}
\title{gseGO}
\usage{
gseGO(
  geneList,
  ont = "BP",
  OrgDb,
  keyType = "ENTREZID",
  exponent = 1,
  minGSSize = 10,
  maxGSSize = 500,
  eps = 1e-10,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  verbose = TRUE,
  nPerm = 1000,
  method = "multilevel",
  adaptive = FALSE,
  minPerm = 101,
  maxPerm = 1e+05,
  pvalThreshold = 0.1,
  ...
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{ont}{one of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{OrgDb}{OrgDb}

\item{keyType}{keytype of gene}

\item{exponent}{weight of each step}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{eps}{boundary for calculating the p value in multilevel mode}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{nPerm}{The number of permutations.}

\item{method}{method of calculating the pvalue, one of "multilevel", "permute" and "sample"}

\item{adaptive}{logical, whether to use adaptive method for calculating pvalue}

\item{minPerm}{minimal number of permutations for adaptive method}

\item{maxPerm}{maximal number of permutations for adaptive method}

\item{pvalThreshold}{pvalue threshold for adaptive method}

\item{...}{other parameters passed to \code{enrichit::gsea_gson()}}
}
\value{
gseaResult object
}
\description{
Gene Set Enrichment Analysis of Gene Ontology
}
\author{
Yu Guangchuang
}
