% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseGO}
\alias{gseGO}
\title{gseGO}
\usage{
gseGO(geneList, ont = "BP", OrgDb, keytype = "ENTREZID", exponent = 1,
  nPerm = 1000, minGSSize = 10, pvalueCutoff = 0.05,
  pAdjustMethod = "BH", verbose = TRUE, seed = FALSE)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{ont}{one of "BP", "MF", "CC" or "GO"}

\item{OrgDb}{OrgDb}

\item{keytype}{keytype of gene}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{seed}{logical}
}
\value{
gseaResult object
}
\description{
Gene Set Enrichment Analysis of Gene Ontology
}
\author{
Yu Guangchuang
}

