% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseKEGG}
\alias{gseKEGG}
\title{gseKEGG}
\usage{
gseKEGG(
  geneList,
  organism = "hsa",
  keyType = "kegg",
  exponent = 1,
  minGSSize = 10,
  maxGSSize = 500,
  eps = 1e-10,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  verbose = TRUE,
  use_internal_data = FALSE,
  seed = FALSE,
  by = "fgsea",
  ...
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{organism}{a supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html',
or a GSON object.}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{exponent}{weight of each step}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{eps}{This parameter sets the boundary for calculating the p value.}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}

\item{seed}{logical}

\item{by}{one of 'fgsea' or 'DOSE'}

\item{...}{other parameter}
}
\value{
gseaResult object
}
\description{
Gene Set Enrichment Analysis of KEGG
}
\examples{
\dontrun{
  data(geneList, package='DOSE')
  gsekegg <- gseKEGG(geneList)
  head(gsekegg)
  kk <- gson_KEGG('hsa')
  gsekegg2 <- gseKEGG(de, organism = kk)
  head(gsekegg2)
}
}
\author{
Yu Guangchuang
}
