% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseKEGG}
\alias{gseKEGG}
\title{gseKEGG}
\usage{
gseKEGG(geneList, species = "hsa", exponent = 1, nPerm = 1000,
  minGSSize = 10, pvalueCutoff = 0.05, pAdjustMethod = "BH",
  verbose = TRUE, seed = FALSE)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{species}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{seed}{logical}
}
\value{
gseaResult object
}
\description{
Gene Set Enrichment Analysis of KEGG
}
\author{
Yu Guangchuang
}

