% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseMKEGG}
\alias{gseMKEGG}
\title{gseMKEGG}
\usage{
gseMKEGG(geneList, organism = "human", exponent = 1, nPerm = 1000,
  minGSSize = 10, pvalueCutoff = 0.05, pAdjustMethod = "BH",
  verbose = TRUE)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{organism}{all KEGG Module supported organisms}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}
}
\value{
gseaResult object
}
\description{
Gene Set Enrichment Analysis of KEGG Module
}
\author{
Yu Guangchuang
}

