% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiPathways.R
\name{gseWP}
\alias{gseWP}
\title{gseWP}
\usage{
gseWP(geneList, organism, ...)
}
\arguments{
\item{geneList}{ranked gene list}

\item{organism}{supported organisms, which can be accessed via the get_wp_organisms() function}

\item{...}{additional parameters, see also the parameters supported by the GSEA() function}
}
\value{
A \code{gseaResult} instance
}
\description{
GSEA analysis for WikiPathways
}
\details{
This function performs GSEA using WikiPathways
}
\author{
Guangchuang Yu
}
