\name{HyperG}
\alias{HyperG}
\title{hypergeometric test}
\usage{
  HyperG(n)
}
\arguments{
  \item{n}{a vector of White balls drawn, White balls Black
  balls balls drawn}
}
\value{
  pvalue
}
\description{
  hypergeometric test for enrichment analysis
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

