% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitr.R
\name{idType}
\alias{idType}
\title{idType}
\usage{
idType(OrgDb = "org.Hs.eg.db")
}
\arguments{
\item{OrgDb}{annotation db}
}
\value{
character vector
}
\description{
list ID types supported by annoDb
}
\author{
Guangchuang Yu
}
