% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_score.R
\name{kappa_score}
\alias{kappa_score}
\title{similarity between categories - kappa score}
\usage{
kappa_score(value)
}
\arguments{
\item{value}{a data.table with 2 columns (two categories) and many rows (objects)}
}
\description{
Function will calculate kappa score between two categories describing a set of elements (GO terms, KEGG pathways, genes).

Function takes a data.table with 2 columns describing to two categories to which each element in rows can belong to. Os and 1s in the table signify membership.

the function is defined as described in this article: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2375021/figure/F2/

relies on data.table for subsetting and vcd package for calculating kappa score given 2x2 incidence matrix

the function is faster than irr::kappa2(ratings = value, weight = "unweighted")

the function is ~25 times slower than dist()

for full gene to GO mapping table (when parent term inherits all annotations of children terms) blows up the memory (over 50GB)
}
\seealso{
\code{\link{categ_dist}}
}
\author{
Vitalii Kleshchevnikov
}
