\name{path2Name}
\alias{path2Name}
\title{convert KEGG pathway ID to pathway Name}
\usage{path2Name(pathIDs)}
\description{provide a vector of KEGG pathway IDs, this function will convert them to corresponding KEGG pathway Names}
\value{KEGG pathway names}
\author{Guangchuang Yu \url{http://ygc.name}}
\arguments{\item{pathIDs}{KEGG pathway IDs}}
