\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,compareClusterResult,ANY-method}
\alias{plot,enrichGOResult,ANY-method}
\alias{plot,enrichKEGGResult,ANY-method}
\alias{plot,groupGOResult,ANY-method}
\title{plot method}
\arguments{
  \item{...}{Additional argument list}
}
\value{
  plot
}
\description{
  plot method generics
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

