% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/compareCluster.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,compareClusterResult,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{compareClusterResult,ANY}(x, type = "dot",
  colorBy = "p.adjust", showCategory = 5, by = "geneRatio",
  split = NULL, includeAll = TRUE, font.size = 12, title = "")
}
\arguments{
\item{x}{compareClusterResult object}

\item{type}{one of bar or dot}

\item{colorBy}{one of pvalue or p.adjust}

\item{showCategory}{category numbers}

\item{by}{one of geneRatio, Percentage or count}

\item{split}{ONTOLOGY or NULL}

\item{includeAll}{logical}

\item{font.size}{font size}

\item{title}{figure title}

\item{...}{Additional argument list}
}
\value{
plot
}
\description{
plot method generics
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
