\name{plot.categoryNet}
\alias{plot.categoryNet}
\title{plot gene net by categories}
\usage{plot.categoryNet(inputList, categorySize="geneNum", showCategory=5,
    pvalue, foldChange, output="fixed")}
\description{plot function of gene Concept Net.}
\value{plotted igraph object.}
\author{Guangchuang Yu \url{http://ygc.name}}
\arguments{\item{inputList}{a list of gene IDs}
\item{categorySize}{setting category size}
\item{showCategory}{number of categories to plot}
\item{pvalue}{pvalue}
\item{foldChange}{foldChange}
\item{output}{output type}}
