\name{plotting.barplot}
\alias{plotting.barplot}
\title{internal function of barplot}
\usage{plotting.barplot(result, title, font.size=12)}
\description{generate a bar plot}
\details{interal use, not for user.}
\value{ggplot object}
\author{Guangchuang Yu \url{http://ygc.name}}
\arguments{\item{result}{a data frame of enrichment result.}
\item{title}{graph title}
\item{font.size}{font size}}
