% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichit.R, R/enrichplot.R, R/reexports.R,
%   R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{geneID}
\alias{geneInCategory}
\alias{gsfilter}
\alias{setReadable}
\alias{bayes_enrich}
\alias{bayes_summary}
\alias{aggregate_enrichment}
\alias{aggregate_omics}
\alias{harmonize_ids}
\alias{select_features_for_ora}
\alias{cnetplot}
\alias{dotplot}
\alias{emapplot}
\alias{goplot}
\alias{gseaplot}
\alias{heatplot}
\alias{ridgeplot}
\alias{read.gaf}
\alias{read.blast2go}
\alias{buildGOmap}
\alias{\%>\%}
\alias{\%<>\%}
\alias{read.gmt}
\alias{read.gmt.wp}
\alias{arrange}
\alias{filter}
\alias{group_by}
\alias{mutate}
\alias{n}
\alias{rename}
\alias{select}
\alias{slice}
\alias{summarise}
\alias{get_organism}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:arrange]{arrange()}}, \code{\link[dplyr:filter]{filter()}}, \code{\link[dplyr:group_by]{group_by()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:n]{n()}}, \code{\link[dplyr:rename]{rename()}}, \code{\link[dplyr:select]{select()}}, \code{\link[dplyr:slice]{slice()}}, \code{\link[dplyr:summarise]{summarise()}}}

  \item{enrichit}{\code{\link[enrichit:aggregate_enrichment]{aggregate_enrichment()}}, \code{\link[enrichit:aggregate_omics]{aggregate_omics()}}, \code{\link[enrichit:bayes_enrich]{bayes_enrich()}}, \code{\link[enrichit:bayes_summary]{bayes_summary()}}, \code{\link[enrichit:geneID]{geneID()}}, \code{\link[enrichit:geneInCategory]{geneInCategory()}}, \code{\link[enrichit:gsfilter]{gsfilter()}}, \code{\link[enrichit:harmonize_ids]{harmonize_ids()}}, \code{\link[enrichit:select_features_for_ora]{select_features_for_ora()}}, \code{\link[enrichit:setReadable]{setReadable()}}}

  \item{enrichplot}{\code{\link[enrichplot:cnetplot]{cnetplot}}, \code{\link[enrichplot:dotplot]{dotplot()}}, \code{\link[enrichplot:emapplot]{emapplot()}}, \code{\link[enrichplot:goplot]{goplot()}}, \code{\link[enrichplot:gseaplot]{gseaplot()}}, \code{\link[enrichplot:heatplot]{heatplot()}}, \code{\link[enrichplot:ridgeplot]{ridgeplot()}}}

  \item{GOSemSim}{\code{\link[GOSemSim:buildGOmap]{buildGOmap()}}, \code{\link[GOSemSim:get_organism]{get_organism()}}, \code{\link[GOSemSim:read.blast2go]{read.blast2go()}}, \code{\link[GOSemSim:read.gaf]{read.gaf()}}}

  \item{gson}{\code{\link[gson:read.gmt]{read.gmt()}}, \code{\link[gson:read.gmt.wp]{read.gmt.wp()}}}

  \item{magrittr}{\code{\link[magrittr:\%<>\%]{\%<>\%}}, \code{\link[magrittr:\%>\%]{\%>\%}}}
}}

