% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSE.R, R/enrichplot.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{geneID}
\alias{geneInCategory}
\alias{gsfilter}
\alias{setReadable}
\alias{cnetplot}
\alias{dotplot}
\alias{emapplot}
\alias{goplot}
\alias{gseaplot}
\alias{heatplot}
\alias{ridgeplot}
\alias{\%>\%}
\alias{\%<>\%}
\alias{read.gmt}
\alias{read.gmt.wp}
\alias{arrange}
\alias{filter}
\alias{group_by}
\alias{mutate}
\alias{n}
\alias{rename}
\alias{select}
\alias{slice}
\alias{summarise}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DOSE}{\code{\link[DOSE]{geneID}}, \code{\link[DOSE]{geneInCategory}}, \code{\link[DOSE]{gsfilter}}, \code{\link[DOSE]{setReadable}}}

  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{summarise}}}

  \item{enrichplot}{\code{\link[enrichplot]{cnetplot}}, \code{\link[enrichplot]{dotplot}}, \code{\link[enrichplot]{emapplot}}, \code{\link[enrichplot]{goplot}}, \code{\link[enrichplot]{gseaplot}}, \code{\link[enrichplot]{heatplot}}, \code{\link[enrichplot]{ridgeplot}}}

  \item{gson}{\code{\link[gson:read-gmt]{read.gmt}}, \code{\link[gson:read-gmt]{read.gmt.wp}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}
}}

