% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplify.R
\docType{methods}
\name{simplify}
\alias{simplify}
\alias{simplify,compareClusterResult-method}
\alias{simplify,enrichResult-method}
\title{simplify method}
\usage{
\S4method{simplify}{enrichResult}(x, cutoff = 0.7, by = "p.adjust",
  select_fun = min, measure = "Rel")

\S4method{simplify}{compareClusterResult}(x, cutoff = 0.7, by = "p.adjust",
  select_fun = min, measure = "Rel")
}
\arguments{
\item{x}{output of enrichGO}

\item{cutoff}{similarity cutoff}

\item{by}{feature to select representative term, selected by 'select_fun' function}

\item{select_fun}{function to select feature passed by 'by' parameter}

\item{measure}{method to measure similarity}
}
\value{
updated enrichResult object

updated compareClusterResult object
}
\description{
simplify output from enrichGO by removing redundancy of enriched GO terms

simplify output from compareCluster by removing redundancy of enriched GO terms
}
\author{
Guangchuang Yu

Guangchuang Yu
}
\references{
issue #28
\url{https://github.com/GuangchuangYu/clusterProfiler/issues/28}
}

