\docType{methods}
\name{summary}
\alias{summary}
\title{summary method}
\usage{
\S4method{summary}{compareClusterResult}(object)
}
\arguments{
  \item{object}{A \code{compareClusterResult} instance.}
}
\value{
A data frame
}
\description{
summary method for \code{compareClusterResult} instance
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

