% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGG.R
\name{viewKEGG}
\alias{viewKEGG}
\title{viewKEGG function is for visualize KEGG pathways
works with enrichResult object to visualize enriched KEGG pathway}
\usage{
viewKEGG(obj, pathwayID, foldChange, color.low = "green",
  color.high = "red", kegg.native = TRUE,
  out.suffix = "clusterProfiler")
}
\arguments{
\item{obj}{enrichResult object}

\item{pathwayID}{pathway ID or index}

\item{foldChange}{fold change values}

\item{color.low}{color of low foldChange genes}

\item{color.high}{color of high foldChange genes}

\item{kegg.native}{logical}

\item{out.suffix}{suffix of output file}
}
\description{
viewKEGG function is for visualize KEGG pathways
works with enrichResult object to visualize enriched KEGG pathway
}
\references{
Luo et al. (2013) Pathview: an R/Bioconductor package for
pathway-based data integration and visualization. \emph{Bioinformatics} (Oxford,
England), 29:14 1830--1831, 2013. ISSN 1367-4803
\url{http://bioinformatics.oxfordjournals.org/content/abstract/29/14/1830.abstract}
PMID: 23740750
}
